There was a lot added to this commit.
Docker compose was created to allow for multiple containers to be ran running the front-end, back-end, and the mongodb that is now in current use for storing submitted espanso triggers.

The API was created and connected to the MongoDB, a new component was also created called TriggerCard, which fetches all Triggers from teh database and puts them at the bottom of the screen to be reviewed and referenced.